// src/types/broadcastx.tsx
export type PostType = {
  id: string;
  url: string;
  aspectRatio: number;
  htmlData: string;
  thumbnail: string;
  type: string;
  uploadTime: number;
};

// {
//   id: '3234',
//   text: 'IBAIS HARDWARE GROUP3',
//   views: 3463,
//   uploadTime: 1696243535636,
//   image: 'https://pbs.twimg.com/media/EkxqUtrXgAEvdJf.jpg:large',
//   aspectRatio: 1,
// }, image
// '<div><b>Specialist in Decorative Products</b><b>MAYUR LAMINATES (MUMBAI) </b><n>Contact: Mr. Jitendra Karia - <a href="98673 45864">98673 45864</a></n><n>For enquiry via WhatsApp click: </n><a href="http://ibais.biz/9821345864" >http://ibais.biz/9821345864</a><n>Email:  jeetukaria4@gmail.com </n><n>Website: </n><a href="https://docowud.in/" >https://docowud.in/</a></div>'
// export type Post = {
//   url: 'https://pbs.twimg.com/media/EkxqUtrXgAEvdJf.jpg:large';
//   aspectRatio: 1;
//   htmlData: '<div><b>Specialist in Decorative Products</b><b>MAYUR LAMINATES (MUMBAI) </b><n>Contact: Mr. Jitendra Karia - <a href="98673 45864">98673 45864</a></n><n>For enquiry via WhatsApp click: </n><a href="http://ibais.biz/9821345864" >http://ibais.biz/9821345864</a><n>Email:  jeetukaria4@gmail.com </n><n>Website: </n><a href="https://docowud.in/" >https://docowud.in/</a></div>';
//   thumbnail: 'https://pbs.twimg.com/media/EkxqUtrXgAEvdJf.jpg:large';
//   type: 'image';
//   uploadTime: 1696243535636;
// };
// url: string;
// aspectRatio: 1
// htmlData: <div><b>Specialist in Decorative Products</b><b>MAYUR LAMINATES (MUMBAI) </b><n>Contact: Mr. Jitendra Karia - <a href="98673 45864">98673 45864</a></n><n>For enquiry via WhatsApp click: </n><a href="http://ibais.biz/9821345864" >http://ibais.biz/9821345864</a><n>Email:  jeetukaria4@gmail.com </n><n>Website: </n><a href="https://docowud.in/" >https://docowud.in/</a></div>
// thumbnail: https://pbs.twimg.com/media/EkxqUtrXgAEvdJf.jpg:large
// type: image;
// uploadTime: 1696243535636;
// 1698969135636